requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
paho-mqtt==2.1.0
pydantic==2.8.2
